Move the exit code test case to E2E file.
